Bug 1268642 - Move datetimeformat jinja helpers to utils
